+++ service/secretsmanager/api.go
@@ -184,6 +184,10 @@ func (c *SecretsManager) CreateSecretRequest(input *CreateSecretInput) (req *req
 // protected secret data and the important information needed to manage the
 // secret.
 //
+// For secrets that use managed rotation, you need to create the secret through
+// the managing service. For more information, see Secrets Manager secrets managed
+// by other Amazon Web Services services (https://docs.aws.amazon.com/secretsmanager/latest/userguide/service-linked-secrets.html).
+//
 // For information about creating a secret in the console, see Create a secret
 // (https://docs.aws.amazon.com/secretsmanager/latest/userguide/manage_create-basic-secret.html).
 //
@@ -1949,42 +1953,20 @@ func (c *SecretsManager) RotateSecretRequest(input *RotateSecretInput) (req *req
 // RotateSecret API operation for AWS Secrets Manager.
 //
 // Configures and starts the asynchronous process of rotating the secret. For
-// more information about rotation, see Rotate secrets (https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotating-secrets.html).
-//
-// If you include the configuration parameters, the operation sets the values
-// for the secret and then immediately starts a rotation. If you don't include
-// the configuration parameters, the operation starts a rotation with the values
-// already stored in the secret.
-//
-// For database credentials you want to rotate, for Secrets Manager to be able
-// to rotate the secret, you must make sure the secret value is in the JSON
-// structure of a database secret (https://docs.aws.amazon.com/secretsmanager/latest/userguide/reference_secret_json_structure.html).
-// In particular, if you want to use the alternating users strategy (https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotating-secrets_strategies.html#rotating-secrets-two-users),
-// your secret must contain the ARN of a superuser secret.
-//
-// To configure rotation, you also need the ARN of an Amazon Web Services Lambda
-// function and the schedule for the rotation. The Lambda rotation function
-// creates a new version of the secret and creates or updates the credentials
-// on the database or service to match. After testing the new credentials, the
-// function marks the new secret version with the staging label AWSCURRENT.
-// Then anyone who retrieves the secret gets the new version. For more information,
-// see How rotation works (https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotate-secrets_how.html).
-//
-// You can create the Lambda rotation function based on the rotation function
-// templates (https://docs.aws.amazon.com/secretsmanager/latest/userguide/reference_available-rotation-templates.html)
-// that Secrets Manager provides. Choose a template that matches your Rotation
-// strategy (https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotating-secrets_strategies.html).
+// information about rotation, see Rotate secrets (https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotating-secrets.html)
+// in the Secrets Manager User Guide. If you include the configuration parameters,
+// the operation sets the values for the secret and then immediately starts
+// a rotation. If you don't include the configuration parameters, the operation
+// starts a rotation with the values already stored in the secret.
 //
 // When rotation is successful, the AWSPENDING staging label might be attached
 // to the same version as the AWSCURRENT version, or it might not be attached
 // to any version. If the AWSPENDING staging label is present but not attached
 // to the same version as AWSCURRENT, then any later invocation of RotateSecret
 // assumes that a previous rotation request is still in progress and returns
-// an error.
-//
-// When rotation is unsuccessful, the AWSPENDING staging label might be attached
-// to an empty secret version. For more information, see Troubleshoot rotation
-// (https://docs.aws.amazon.com/secretsmanager/latest/userguide/troubleshoot_rotation.html)
+// an error. When rotation is unsuccessful, the AWSPENDING staging label might
+// be attached to an empty secret version. For more information, see Troubleshoot
+// rotation (https://docs.aws.amazon.com/secretsmanager/latest/userguide/troubleshoot_rotation.html)
 // in the Secrets Manager User Guide.
 //
 // Secrets Manager generates a CloudTrail log entry when you call this action.
@@ -2468,6 +2450,10 @@ func (c *SecretsManager) UpdateSecretRequest(input *UpdateSecretInput) (req *req
 // To change the secret value, you can also use PutSecretValue.
 //
 // To change the rotation configuration of a secret, use RotateSecret instead.
+//
+// To change a secret so that it is managed by another service, you need to
+// recreate the secret in that service. See Secrets Manager secrets managed
+// by other Amazon Web Services services (https://docs.aws.amazon.com/secretsmanager/latest/userguide/service-linked-secrets.html).
 //
 // We recommend you avoid calling UpdateSecret at a sustained rate of more than
 // once every 10 minutes. When you call UpdateSecret to update the secret value,
@@ -6174,8 +6160,9 @@ type RotateSecretInput struct {
 	// Specifies whether to rotate the secret immediately or wait until the next
 	// scheduled rotation window. The rotation schedule is defined in RotateSecretRequest$RotationRules.
 	//
-	// If you don't immediately rotate the secret, Secrets Manager tests the rotation
-	// configuration by running the testSecret step (https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotate-secrets_how.html)
+	// For secrets that use a Lambda rotation function to rotate, if you don't immediately
+	// rotate the secret, Secrets Manager tests the rotation configuration by running
+	// the testSecret step (https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotate-secrets_how.html)
 	// of the Lambda rotation function. The test creates an AWSPENDING version of
 	// the secret and then removes it.
 	//
@@ -6183,7 +6170,12 @@ type RotateSecretInput struct {
 	// the secret immediately.
 	RotateImmediately *bool `type:"boolean"`
 
-	// The ARN of the Lambda rotation function that can rotate the secret.
+	// For secrets that use a Lambda rotation function to rotate, the ARN of the
+	// Lambda rotation function.
+	//
+	// For secrets that use managed rotation, omit this field. For more information,
+	// see Managed rotation (https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotate-secrets_managed.html)
+	// in the Secrets Manager User Guide.
 	RotationLambdaARN *string `type:"string"`
 
 	// A structure that defines the rotation configuration for this secret.